+++ python/ray/autoscaler/v2/instance_manager/node_provider.py
@@ -5,7 +5,7 @@ from typing import Dict, List, Set
 from ray.autoscaler._private.node_launcher import BaseNodeLauncher
 from ray.autoscaler.node_provider import NodeProvider as NodeProviderV1
 from ray.autoscaler.tags import TAG_RAY_USER_NODE_TYPE
-from ray.autoscaler.v2.config import AutoscalingConfig
+from ray.autoscaler.v2.instance_manager.config import AutoscalingConfig
 from ray.core.generated.instance_manager_pb2 import Instance
 
 logger = logging.getLogger(__name__)
