+++ python/ray/autoscaler/v2/instance_manager/ray_installer.py
@@ -4,7 +4,7 @@ import subprocess
 from ray.autoscaler._private.updater import NodeUpdater
 from ray.autoscaler._private.util import with_envs, with_head_node_ip
 from ray.autoscaler.node_provider import NodeProvider as NodeProviderV1
-from ray.autoscaler.v2.config import AutoscalingConfig
+from ray.autoscaler.v2.instance_manager.config import AutoscalingConfig
 from ray.core.generated.instance_manager_pb2 import Instance
 
 logger = logging.getLogger(__name__)
