+++ python/ray/autoscaler/v2/tests/test_ray_installer.py
@@ -7,7 +7,7 @@ import pytest  # noqa
 
 from ray._private.test_utils import load_test_config
 from ray.autoscaler.tags import TAG_RAY_NODE_KIND
-from ray.autoscaler.v2.config import AutoscalingConfig
+from ray.autoscaler.v2.instance_manager.config import AutoscalingConfig
 from ray.autoscaler.v2.instance_manager.ray_installer import RayInstaller
 from ray.core.generated.instance_manager_pb2 import Instance
 from ray.tests.autoscaler_test_utils import MockProcessRunner, MockProvider
